new_fraud_app/requirements.txt
# Core dependencies
azure-identity>=1.12.0
semantic-kernel>=0.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Azure AI dependencies
azure-ai-agents>=1.0.0
azure-ai-generative>=1.0.0b11
azure-core>=1.29.0
openai>=1.0.0
azure-ai-projects>=1.0.0b11

# Additional utilities
python-dateutil>=2.8.2
typing-extensions>=4.5.0